packages/python.org/services/openai/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.115.14
openai==1.99.9
pydantic==2.11.7
python-dotenv==1.0.1
uvicorn==0.34.2
